applets/devicenotifier: create the SpaceMonitor class one per device
Rename the SpaceMonitor class to SpaceInfo and create one for each device.
Removes the demand to have the list of devices. Also move the logic of
two state devices handling from DeviceControl to it as it is only related to
space update.
Also move functions related to space updates from DeviceControl to SpaceInfo