KCMColors: Do not forward the output of editDialogProcess
When we are done editing a colorscheme and save it, dialogProcess
outputs in stdout the names of the colorschemes we save.

ForwardedChannels would send those names to the stdout of the
main process (the KCM), so the dialogProcess stdout was always
empty.

Instead, just use defaults: Separate channels (stdout/stderr) and then
just read it as usual.

This then allows the reloading to work.

BUG: 504007


(cherry picked from commit 6993f8e3b1652ae09ebdfbc51a301cf2f7f07ecb)

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>