applets/notifications: Use a proper component for the speed chart
Children of a Loader are not part of the loader's contents, they're just
children of the loader. To make sure we actually have the chart as
Loader contents, wrap it in an item and set it as sourceComponent. This
ensures we don't try to load the chart when it's not actually visible or
needed.

CCBUG: 503099