applets/notifications: Use a proper component for the speed chart
Children of a Loader are not part of the loader's contents, they're just
children of the loader. To make sure we actually have the chart as
Loader contents, wrap it in an item and set it as sourceComponent. This
ensures we don't try to load the chart when it's not actually visible or
needed.

CCBUG: 503099


(cherry picked from commit 8f81580780253f98210731c8962f33c67de08e98)

Co-authored-by: Arjen Hiemstra <ahiemstra@heimr.nl>