klipper: ignore file previews from network mounts
A remote desktop application like RustDesk that supports seamless clipboard might
leave a local file url in the clipboard but the file is actually a
remote file. If there are many urls like that and those remote files are big
enough, plasmashell will hang in the end.

QUrl::isLocalFile doesn't check if a file is on a network mount.
Instead, use KFileItem::isSlow which has a more comprehensive logic to
detect remote files.