notifications: Allow choosing screen where notification popups show
By default, it follows with a complicated heuristic where the notification
plasmoid is. However, there can be use cases for showing them on a
different screen.

This adds a screen selector in the popup position dialog. When setting
a custom screen, the corner has to be configured since we don't
have a plasmoid to follow (or rather, trying to find one on that new
screen makes the already complex code even worse).

If the configured screen isn't available, the primary one is used
until it becomes available again.

BUG: 446961