kcm_regionandlang: fix locale list item background is not highlighted
BUG: 507451

In locale list, the selected property is set to iconTitle component,
but not to its parent localeDelegate. Thus, the selected item only has
highlighted text, the background is unchanged.
On light theme, the selected item is white on white. Set the highlighted
property of localeDelegate to make selected item render properly.