libdbusmenuqt: Add DBusMenuModel and DBusMenuView
The DBusMenuModel provides a tree-like model representing a menu with
the com.canonical.dbusmenu interface.

The DBusMenuView type provides the corresponding view object. It can
display the toplevel menu (i.e. the one that includes Files, Edit, View,
and so on entries), or the second level menus (e.g. a Files submenu).
The latter mode is more suitable for the global menu applet, where we
would like to reuse the menu for various submenus.