applets/appmenu: Port to DBusMenuModel and DBusMenuView
Currently the global menu applet sends the about to show event for
submenus even though they are not actually shown. This breaks Firefox.

The applet does it for two reasons: to improve the responsiveness of
the global menu when moving the pointer, and to work around some
design flaws in the DBusMenuImporter. The global menu wants to re-use
the menu object for submenus. In order to do that, it moves actions
from QMenu objects owned by the DBusMenuImporter to its own. Sending
opened events acts a way to ensure that the layout stays consistent.

The DBusMenuModel and DBusMenuView types suit better the needs of the
global menu. It also fixes Firefox breaking.

Unfortunately, this change also removes the search capability. The
previous implementation heavily relied on Qt widgets specific details,
they caused problems in the past and they can't work with this design.
The search feature also did not work properly because the appmenu
fetches a shallow menu layout subtree, it doesn't fetch the whole tree.
This applet is the wrong place for such a feature in general, it
should be provided by another component (there is a patch for this
already). Given that being able to search is not the main focus of this
applet, the search capability was dropped.

See also https://bugzilla.mozilla.org/show_bug.cgi?id=1975681.

BUG: 506439