applets/appmenu: Fix wrong menu position
The QtWayland QPA blocks repaints if a menu is repositioned, but it is
not enough to ensure that the window position is correct. For example,
consider the following case:

 - ask the compositor to move a menu to 10,10
 - ask the compositor to move the menu to 20,20
 - apply a configure event that moves the menu to 10,10
 - add an action in the menu
     - this will change the geometry and result in another reposition
       request with 10,10
 - apply a configure event that moves the menu to 20,20
 - apply a configure event that moves the menu to 10,10

This behavior can be observed with the appmenu popup sometimes. In rare
cases, it gets stuck at an old position.

In general, async window position is a very very hard problem, and it is
better to avoid using absolute window coordinates. This change makes the
global menu use the positioner API available in QtWayland. It should make
popup repositioning very robust on Wayland.