[libtaskmanager] Drop taskmanagerrulesrc machinery
We have very complex code for mapping a window to a desktop file.

A huge part of the complexity is the taskmanagerrulesrc file that
allows to define custom mappings (i.e. workarounds).

Realistically no user is going to do that though, since the mechanism
is obscure and undocumented.

It ships with a few standard mapping rules, but all of those seem either
unnecessary or not working as intended.

Dropping this allows for massively simplifying the mapping code