applets/devicenotifier: Allow services to execute without mounting
This adds a new property for services to define in their .desktop files
to skip mounting the device automatically. By default, it will still
mount before executing the action.

This is useful for actions that directly interact with the device on
their own terms, like for partition editing. The applet auto-mounting
the device is useless in that case!