runners: remove group separator from input first
### Reason for the change
In calculator runners, we replace commas with dots.
But in locales use comma as group separator, this change the value of the input.
For example, 1,000 will be changed to 1.000, which is incorrect.
Remove group separator from input first, so the input value remains unchanged.

### Test plan
Compile plasma-workspace and krunner using kde-builder. Launch krunner and type "10,000*2".
If you want to test other locales, first enable and generate the locales (distro specific),
then run krunner with "LC_NUMBERS=es_PE.UTF-8 krunner".

Check if the calculation result is correct.

### Screenshots or screen recordings
Before:
![Screenshot_20250907_130658_before](/uploads/a40c509db00f09c0a8e00b766e833add/Screenshot_20250907_130658_before.png)
After:
![Screenshot_20250907_130624_after](/uploads/21098e34fa4f0b83a9a68658dbbbe1c1/Screenshot_20250907_130624_after.png)