krunner: Prevent window of results overflowing screen vertically when floating (positioned at centre of screen)
<!-- Thank you for submitting a merge request to make KDE Software better! Use the information from the commit message(s) to fill in the template below.

### Original commit message(s)

* krunner: Prevent window overflowing screen vertically when floating

(For future reference, follow the guidelines at https://cbea.ms/git-commit/#seven-rules)
-->


### Reason for the change

Fixes a UI bug which also gets in the way of functionality: the lowest KRunner search results are inaccessible when the position on screen of KRunner is centre ("floating") rather than top, because the results list extends below the screen's boundary.

### Test plan

<!--
Extend any related autotests with new test cases to validate your changes. If none exist, consider adding one if you have the technical skills to do so.

Otherwise, at least describe how you tested the change, and how someone reviewing this merge request can test it themselves. No need to mention obvious things like "Apply the changes".
-->

**There aren't any autotests for the height, for position on screen = top *or* centre, but this may be a good learning opportunity for me** (I already know Python with classes); I'll let you know if I've managed to add them around tomorrow midday 2025-09-04T11:00:00Z.

#### Steps to reproduce:
1. Alt+Space to open Krunner; click the "configure krunner" icon to the left of the search bar, and wait for a configuration window to open.
2. **Set "position on screen" to "centre"** (termed floating in the code) and apply the config
3. Type something that gets a lot of results. For me "time" or "tim" works.
4. If this doesn't fill the whole screen height, lower the screen resolution in system settings then repeat the instructions.

#### Expected:
Bottom of results list is exactly at bottom of screen, and scrollbar is present. This is already the observed result when "position on screen" is "top".

#### Observed:
Bottom of results list extends below the bottom of the screen, and scrollbar can be present but this still doesn't allow seeing the lowest results on the list.

### Screenshots or screen recordings

<!--
If this merge request introduces a visual change, please add before-and-after screenshots in the following format:

| Before | After |
| ------ | ----- |
| [drag "before" screenshot here] | [drag "after" screenshot here] |
-->

(Zoom in to the bottom of each image)

| Before | After |
| ------ | ----- |
| ![Not floating (centre-aligned): correct. Floating (centre-aligned): bug: overflows screen. Please read observed section above.](/uploads/a22693db16dabc940087099a8fe3b755/image.png){width=969 height=1245} | ![Not floating (centre-aligned): correct. Floating (centre-aligned): correct. Please read expected section above.](/uploads/0ecff87c2ccc07c5778644ba8c1f79a0/image.png){width=962 height=1092} |

### Bugs fixed

<!--
If the changes in this merge request fix any Bugzilla tickets, add the following keyword for each one:

BUG: [number of the bug report]
-->

Bugfix for a bug **not** on bugs.kde.org (sorry, I didn't register it there as it was quicker to fix straight away; should I add it now and reference it from the MR for the record?).