Make all panels reserve space when in edit mode
Currently, only the currently edited panel always reserves space,
regardless of visibility and number of panels on the same side.
This makes it impossible to enter edit mode for panels that are
overlapping other ones. Instead, we make all panels always reserve
space, so that they can always be interacted with.