DeviceNotifications: Cooldown before notifications
In some cases, for example having faulty monitor or
faulty wire, the monitor can inform itself to be
plugged/unplugged multiple times.

In these cases, we should check if the output
was recently added/removed, and if so, ignore
sending notifications: Wait for 2000ms before
allowing for the added/removed display to
create a notification.

Otherwise the notification will be sent as usual.

BUG: 494330