applets/mediacontroller: Workaround for common crash in QQuickStackView
This commit adds a workaround for QTBUG-140018 where a regression in Qt 6.9.2 causes a segfault as qmlEngine(view) returns null in QQuickStackElement::initialize, because a parent loader set the engine to null when unloading, but we still run.

We avoid this by not pushing a new Image to the StackView when it isn't visible, which is coincident with the above and the cause of the crash.

This does not regress any behaviour, functionality or animation.

BUG: 509192

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>