ksmserver: Fix memory management of notification
Cleanup of notifications happens automatically when they are closed.

There is a fallback timer here that uses the state to check if it should
perform additional cleanup.

This is not safe as if we ever had two notifications within 5 seconds we
could be in the state WaitingForKnotify but still have the first
notification active at this point.

We can let KNotification autodelete continue to take care of it.

BUG: 508726


(cherry picked from commit 3f2b85e73fa104798b333296f241f0a4aee3702d)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>