ContanmentLayoutManager: Limit relayout of placeholder
While dragging an applet around, the placeholder gets laid out at the
position an applet would be normally positioned to show the final
position and size the applet would have.
But this is a somewhat expensive operation, which is not a problem
with devices with a relatively low polling rate, but becomes
a problem with gaming mice that can have a refresh rate of 1000Hz
and up.

Limit the relayout to 10 fps, since it moves snapping, we don't
need to be smooth at all, just as cheap as possible
BUG:509860