GridLayoutManager: Fix some freezes that would happen with dragging widgets
GridLayoutManager does a lot of calculations when
looking for the next fitting spot for a widget.
These are while loops that can often get stuck
trying to find something perfect.

Instead, try only searching for small areas:
We lose some precision for available spots, but
we gain more performance. This is especially noticeable
on high resolution displays.

We should also not start calculating for the
position from start, but from the position
of the item, otherwise we'll do a lot of work
for minimal precision gains.

CCBUG: 509860