plasma-shutdown: Exit after manual session save
plasma-shutdwon is a DBus activated executable that is meant to run one
task and exit swiftly.

The call to manually save a session left the process idling. This was
problematic as the logout greeter checked the presence of this process
to see if an existing shutdown is in progress

BUG: 488998


(cherry picked from commit e1b251fa874e5bb1dbb009440724621a462cd691)

Co-authored-by: David Edmundson <kde@davidedmundson.co.uk>