OsdItem: reverse icon if possible when using an RTL language
The layout here is very strongly directional. That means when the layout
reverse (i.e. when using an RTL language, the icon in the layout needs
to reverse too if it has any strongly directional elements. This commit
does that.

There are no visual changes when using an LTR language or when the
current icon theme does not have RTL versions of the icon used here.

CCBUG: 509598