wallpapers/image: Improve kcm delegate sizing for mobile
Currently the wallpaper grid is basically unusable on mobile due to the
delegates being too small. The calculations were determined based on
dividing the screen size, which causes issues on phones with smaller
resolutions. This changes the calculation to use unit factors instead, with
an intended delegate length (larger for mobile for touch reasons). This
results in less delegates being added to each row.

This accompanies
https://invent.kde.org/frameworks/kcmutils/-/merge_requests/277, which
gets rid of mobile-specific sizing tweaks to the KCM GridView, so that
the delegates can reflect the screen dimensions.