wallpapers/image: swap first image on duplicate
When reaching the end of the randomized image slideshow,  
don't skip the first item after shuffling the randomOrder if it's  
the same as the previous one.

Instead, swap the item at index 0 with a random item at index 1..n-1,  
so that no items are skipped and the slideshow size remains the same.

BUG: 509711
FIXED-IN: 6.6.0