[libnotificationmanager] Implement ActionLabelsRole in JobsModel
Currently we don't handle this role at all and return QVariant.

This triggers https://bugreports.qt.io/browse/QTBUG-140928 which
causes hasSomeActions to be true for job finish notifications,
causing them to not be closed on expire.

By returning a data object of the right type we work around the
issue

BUG: 510244