Klipper: Allow match patterns and commands to be disabled by a check box
A feature request from a long time ago.  Useful if the user wants a
match pattern or command to temporarily not apply, but may need to
reactivate it in the future without having to go through the complicated
process of adding it from new again.

The configuration for saving an enabled state for each defined command
already existed, the only changes needed were the configuration GUI and
actually taking note of the enabled state when creating the action
popup menu.

BUG: 7952