scripting: Fix setting resolution independent panel configuration
Certain configuration keys are treated as resolution independent and
PanelView::restore() uses config().parent() for reading/writing them.
Mirror that in the scripting wrapper for panels so that the configuration
is written in the right place for the PanelView to pick it up.


(cherry picked from commit ce2d9bf05d1b195c8bc8562ed1b94d168214f119)

Co-authored-by: Fabian Vogt <fabian@ritter-vogt.de>