RFC: Use simpler animation driver from Qt
The way Qt currently works (qsgcontext.cpp) is the default timer (QSGDefaultAnimationDriver) is driven when the window repaints. This breaks for occluded windows or multiple windows, so there's a backup mechanism QSgThreadedRenderLoop::m_animation_driver that backs things via a timer.

Given we know we'll always have multiple windows we're not hitting this optimised paths

As per https://doc.qt.io/qt-6/qtquick-visualcanvas-scenegraph.html we have some other options.

Note: Starting from Qt 6.5, the threaded render loop offers the possibility of opting in to another animation driver, based solely on the elapsed time (QElapsedTimer). To enable this, set the QSG_USE_SIMPLE_ANIMATION_DRIVER environment variable to a non-zero value. This has the benefits of not needing any of the infrastructure for falling back to a QTimer when there are multiple windows, not needing heuristics trying determine if vsync-based throttling is missing or broken, being compatible with any kind of temporal drifts in vsync throttling, and not being tied to the primary screen's refresh rate, thus potentially working better in multi-screen setups. It also drives render thread animations (the Animator types) correctly even if vsync-based throttling is broken or disabled. On the other hand, animations may be perceived as less smooth with this approach. With compatibility in mind, it is offered as an opt-in feature at the moment.

This seems to be a direction Qt wants to push for and it simplifies things a lot.

If we can then unlock the use of Animators we can use them to get the max framerate possible.

Note this env leaks into other processes currently.
Frankly this isn't a bad thing, but if we do want that maybe it should be in startplasma