wallpapers: rephrase color-aware wallpaper option to be accurate
In the past it reflected the lightness or darkness of the color scheme,
but this was changed in d678ab7f692875d66ae74813e08dffd72965b76d to
reflect the Plasma style. Update the text to match the implementation.


(cherry picked from commit ad131736240c2e7bcc132b3c12275df45a8c4609)

Co-authored-by: Nate Graham <nate@kde.org>