wallpapers/image: Remove Tiled mode
We would like to improve wallpaper image memory usage. Unfortunately,
this means that we need to re-implement image item bits, including all
the various fill modes.

Unfortunately, the `Tile` mode cannot be re-implemented with vanilla
QtQuick APIs both for RHI and software rendering. However, this also
raises questions whether the tiled mode has use cases in 2025, and
whether it needs to be included with mainstream image plugin.

The tiled mode has been re-implemented as a separate plugin in kdeplasma-addons.