shutdown: Stop graphical-session.target directly
The Plasma session has a hierachy of targets that requires the lower
target with StopWhenUnneeded=yes.
On logout we would stop plasma-workspace.target and systemd would unravel
the now unneeded targets. This scheme fails as soon as something explicitely
depends on a lower target for example by having
Requsite=graphical-session.target
as now the target is still needed and graphical-session.target is not stopped
and only the plasma-workspace, plasma-workspace-wayland/x11 and plasma-core
targets are stopped breaking log out.