wallpapers/image: Avoid reloading image model when target size changes
If the screen resolution changes, the image model will be reloaded. On
the other hand, the wallpaper list is independent of the screen
resolution. The model is reloaded so the package list model looks up new
preferred images.

It's also worth noting that the wallpaper image plugin has been slowly
but steadily offloading some of its logic to the view side, which is a
very good idea for code complexity management! If the target size
changes, then the `Image` or the `TransientImage` will lookup the right
image in a package or reload an image with a new scaled down size
without any involvement from the model.

This helps with the startup of the dialog on wayland a little bit, the
model will be loaded only once now. But in long term, the wallpaper
dialog needs to be told what output it must target, not follow the
output of the container window. The startup of the wallpaper dialog can
be further improved by using `TransientImage`, so previews are loaded
once too.