kcms/region_language: Use correct JEDEC unit in binary dialect description
There is a discrepancy between the JEDEC unit as expressed by KFormat ('1.0 KB'), and as shown in the description hint in the KCM ('1.0 kB').

According to JESD100B.01, which defines "kilo (K) (as a prefix to units of semiconductor storage capacity)", the correct is '1.0 KB'.

As such, the discrepancy is resolved by changing the KCM binary dialect description.