applets/notifications: Fix notification time labels
This was introduced in 9f9b96e1e443d8f4c5636ef99976674267c508d3 it
seems, due to the fact that the logic which decides which timestamp to
use (created or updated) depends on the property being nullable.
However, when we don't use var but a specified property type it doesn't
become nullable - just a regular invalid JS Date object.

We can use isNaN here while keeping the existing required property
intact and that fixes the issue.

BUG: 511908
FIXED-IN: 6.5.3