applets/notifications: A little bit more cleanup of ||
This is a follow-up to the change to required model properties, where
these are mostly functionally useless now. Which isn't a bad thing, per
se but lets remove the dead code.

I made sure to keep the ones that *are* actually being used, such as
popup.image and popup.iconName. The rest of them are either default
values (NormalUrgency) or empty/invalid ones which are fine. It doesn't
seem to obviously regress anything on my end.