applets/systemtray: Consolidate config pages
Right now it has two pages with contents that want different amount of
space. As a result, there's no good window size that doesn't make one of
the pages look ridiculous. The Entries page is also slightly hidden.

To alleviate this, merge the two pages together. Now there's one, it
wants to be large, it looks good when large, and everything is
discoverable.