applets/systemtray: remove inline header and move InlineViewHeader inline
The current inline header doesn't offer much value; it's pretty obvious
what the columns in the table are all about, with maybe the exception of
the shortcut column. This could be improved by giving the empty shortcut
buttons better icons and text.

Remove it, and allow the InlineViewHeader item to live where it was
designed to: inline, as the view's header.