klipper/declarative: allow Return/Enter from tab buttons
Klipper allows you to press Enter/Return from the search field to
directly copy the currently highlighted item and close the window.
This does not work if the user is on the tab bar, so when switching
from All History to Starred, the user needs to press down first before
they can press Enter to set the clipboard - the keys by themselves do
nothing. This is particularly annoying as neither the TabButtons nor
list delegates properly indicate that they have activeFocus, so there
is no visual difference between the two states (which should also be
fixed, but is a much larger issue).

Instead, we allow pressing Enter/Return from the tab buttons as well,
reusing the implementation for the search field. This makes it simpler
and less confusing to switch to the most recently used starred item from
the All History view (the other direction in principle as well, but that
is less useful as the top item there is also the current clipboard entry
so there's little reason to activate that one).