shell: Harden PanelRulerView against screen changing due to geometry updates
This change re-arranges the code so it should be harder for the panel
ruler and the panel view end up on different outputs. In hindsight, it
will be helpful to have setDesiredOutput() in layer-shell-qt to be sure
that the ruler will be on the correct output.

BUG: 482916
(cherry picked from commit 7b99e9c12ead02aac38b5ba8ec7cf9a2b403a023)