libkmpris/kded: add global shortcuts for seeking
Plasma ships with global shortcuts for most common player control
actions. The exception is seeking, which is supported by MPRIS, but
which we do not expose as a potential shortcut. This can be convenient
e.g. when listening to audio books/podcasts or while working with
transcripted audio.

This change adds short (5s) and long (30s) skips forward and backward
as user-bindable shortcuts with no default binding.

FEATURE: 416130
FIXED-IN: 6.6.0