klipper: Introduce  a separate proccess to own the clipboard
plasmashell owning the selection directly has lead to a multitude
of problems, need to be careful to not deadlock itself complicated
code to handle transfer clipboard from data control to qt that doesn't
work correctly in all cases, etc. Instead use a separate process
to own the clipboard so we dont need to worry about such problems anymore
which can just read the data out of the database which makes this
quite convenient as we only need to pass the uuid to it. Existing
code in HistoryModel will detect that the data is the same as a
entry in it so no duplicate entries or infinite loops happen.