applets/kicker: set mirroring for submenu opening explicitly
Kicker's submenu placement is based on the system's layout direction.
This mostly works, but breaks for far-edge panels (right side in ltr,
left side in rtl), where Kicker (partially) flips its orientation. This
isn't a huge problem for submenus (which have their position forced by
the screen edge anyway), but sub-submenus open in the wrong direction,
covering Kicker's (main) menu.

Instead, we make the direction where dialogs should be placed an
explicit property of the SubMenu class, so we can set it from the qml
side.


Co-authored-by: Méven Car <meven@kde.org>