wallpapers/image: Fix wallpaper previews not honoring aspect ratio for sourceSize
If the minimum source size is forced (it happens with my 4K monitor
scaled at 130%), then the sourceSize may have an aspect ratio of 1:1.

This can result in the portait orientation image being loaded instead of
the landscape one.

This change makes sure that the minimum forced source size follows the
aspect ratio of the screen.