applets/kicker: be more assertive about not having children
Kicker's AbstractModel has IsParent and HasChildren roles, which are
inherited by RecentUsageModel. That model, however, does not return
bool values for these roles, only an emoty QVariant. This is not a
problem generally, but does cause issues if a submenu dialog is reused
in the kicker frontend - it may already have these values set from a
previously open submenu, and the empty QVariant from the model does not
override this, so recent apps or files may incorrectly try to open
subsubmenus if the user switches from a submenu with subgroups directly
to the favorites.

Instead, explicitly return false for these; entries in the favorites
menu are not groups and do not have children.