startkde: Apply day-night global theme earlier
The day-night global theme is applied by kded, which may happen some
time later, for example, when the splash screen loads, etc. This makes
the startup process look less "smoother". It is typically seen in the
morning of the next day.

This change moves the day-night theme application earlier in the startup
process, so when the remaining plasma components are loaded, they will
read the new defaults.