shell/panelview: Fix race condition for panel sizing
We can't have this geometry check here, because there are cases where
the geometry gets updated by other causes outside of
positionAndResizePanel, and so the layershell window dimensions need to
be updated as well.

This can cause the layershell window's dimensions to be out of sync with
the panel contents, causing improper panel sizes. It's mainly triggered
by changing the dimensions of the panel is quick succession.