replace Qt.application.layoutDirection with Application singleton
qmllint fails to properly handle the Qt.application property
(QTBUG-119643). Upstream recommends using the Application singleton
instead, as it resolves issues with the types (which are variable
depending on the imported modules) and is more accessible for tooling
(codereview qtdeclarative/+/350080). The main use for Qt.application
in p-w is to access the layoutDirection.
This change ports p-w to using the Application singleton.