applets/devicenotifier: properly handle fast add/remove signals
Sometimes multiple add and remove signals are emitted, for example when formatting a
storage device. The applet checks whether the device is already present in the device list
and refuses to add it again. This can cause the device to not be shown at all: the device
may be in the process of being removed but still present in the list, and if an add signal
arrives at that moment, it will be ignored.

To handle this correctly, force-remove the device and add it again. Also, do not delay
removal if the message does not exist.