applets/kicker: make non-runnable entries return empty favoriteID
Similar to 510bd9592ed971a6cfb493498b4b8436980c4f95

The appsmodel returns an empty QVariant for the favoriteID role if
the entry is not runnable. This would be fine, but for submenus
Kicker reuses the existing ListView and delegates, and that does
not cause the property value to change. so if a user has a submenu open
that contains apps, then changes to a submenu that has subsubmenus
in that position and right-clicks that submenu, they'll be offered the
option to add it to the favorites (and it'll add whatever app was in the
previous submenu).

Instead, return an empty QString in such cases - this will cause the
property value to change, and kicker does not offer to make it a
favorite anymore.