applets/digital-clock: limit date format example width
The example display for the label format is synced in its size with
some other components making a visual column, and basing it on its
maximum element. This is fine in the regular case, but the user can
make the format string ridiculously long, which makes everything else
push off the side of the page as well, causes weird visual glitches
(such as the page occasionally animating a flip to itself) and is
overall pretty silly.

Instead, we keep the label size synced to the other elements, but only
from the other components to the label, limiting its size, and clipping
it if necessary. We also add a Tooltip to allow users to view their
creation in full.