applets/devicenotifier: don't set status in onCompleted
Devicenotifier tries to set the plasmoid status to passive in
Component.onCompleted if nothing is connected. But the check can't work:
it reads the property deviceCount from the PlasmoidItem, but it doesn't
have one, deviceCount is a property of the model, not the PlasmoidItem.
Reading it will always just give undefined, and that's never exactly
equal to zero.

It also isn't really necessary, as the status is already set in a
binding. We wouldn't want to overwrite that binding, as the applet
should appear and disappear as the user connects or removes devices, not
just fixed depending on the state at the applet startup.

The whole bit does nothing at the moment and there's nothing useful for
it to do, so let's drop it.